fix #17

+++ src/main/java/org/wxjs/les/modules/check/export/TsitecheckExport.java
@@ -233,27 +233,33 @@ public class TsitecheckExport {
 
             /* ============ 表格中的第五部分数据 ============ */
             Image checkerSig=null;
-            try {
-         	   String filename=this.base64StringToImage(this.tsitecheck.getCheckerSig().getSignature());
-         	   checkerSig=Image.getInstance(filename);
-         	   FileUtils.deleteFile(filename);
-            } catch (MalformedURLException e) {
-        	   logger.error("checkerSig error",e);
-           } catch (IOException e) {
-         	   logger.error("checkerSig error",e);
-           }
+            if(this.tsitecheck.getCheckerSig()!=null && !StringUtils.isEmpty(this.tsitecheck.getCheckerSig().getSignature())){
+                try {
+              	   String filename=PdfUtil.base64StringToImage(this.tsitecheck.getCheckerSig().getSignature());
+              	   checkerSig=Image.getInstance(filename);
+              	   FileUtils.deleteFile(filename);
+                 } catch (MalformedURLException e) {
+             	   logger.error("checkerSig error",e);
+                } catch (IOException e) {
+              	   logger.error("checkerSig error",e);
+                }
+            	
+            }
          
            Image partySig=null;
             
+           if(this.tsitecheck.getPartySig()!=null && !StringUtils.isEmpty(this.tsitecheck.getPartySig().getSignature())){
 	           try {
-	     	   String filename=this.base64StringToImage(this.tsitecheck.getPartySig().getSignature());
+	     	   String filename=PdfUtil.base64StringToImage(this.tsitecheck.getPartySig().getSignature());
 	         	   partySig=Image.getInstance(filename);
 	         	   FileUtils.deleteFile(filename);
 	            } catch (MalformedURLException e) {
 	         	   logger.error("partySig error",e);
 	            } catch (IOException e) {
 	         	   logger.error("partySig error",e);
-	            }
+	            }       	   
+           }
+
 	           widths = new float[]{8, 42,8,42};
 	           table = new PdfPTable(widths.length);
 	           table.setWidths(widths);
@@ -301,7 +307,11 @@ public class TsitecheckExport {
     	cell.setVerticalAlignment(Element.ALIGN_MIDDLE); //垂直    	
     	tableSub.addCell(cell); 
     	
-    	phrase = new Phrase(DateUtil.formatDate(date, "yyyy年MM月dd日"), PdfUtil.getFont10(Font.NORMAL));
+    	phrase = new Phrase("");
+    	if(sig != null){
+    		phrase = new Phrase(DateUtil.formatDate(date, "yyyy年MM月dd日"), PdfUtil.getFont10(Font.NORMAL));
+    	}
+    	
 		cell = new PdfPCell(phrase);
     	cell.setBorderWidth(0);
     	cell.setHorizontalAlignment(Element.ALIGN_RIGHT); //水平
@@ -340,21 +350,6 @@ public class TsitecheckExport {
         response.setHeader("Content-Disposition", "attachment;filename=" + Encodes.urlEncode(fileName));
         this.generate(response.getOutputStream());
         return this;
-    }
-    
-    public String base64StringToImage(String base64String) {
-    	String filename=Global.getConfig("userfiles.basedir")+"/"+IdGen.uuid()+".png";
-    	try {
-    		byte[] bytes1=Base64.decodeBase64(base64String);
-        	ByteArrayInputStream bais=new ByteArrayInputStream(bytes1);
-			BufferedImage bi=ImageIO.read(bais);
-			File f1=new File(filename);
-			ImageIO.write(bi, "png", f1);
-		} catch (IOException e) {
-			e.printStackTrace();
-		}
-		return filename;
-    	
     }
     
 }
